safari-600.7-branch/LayoutTests/fast/forms/option-mouseevents-expected.txt

This page tests whether a click event on a list box item propogates with the correct target and positioning values.
Bug 3248: Mouse events on OPTION element seem to be ignored
https://bugs.webkit.org/show_bug.cgi?id=3248





PASS: event type should be mousedown and is
PASS: event target should be [object HTMLOptionElement] and is
PASS: event.pageX should be 22 and is
PASS: event.pageY should be 104 and is
PASS: event.offsetX should be 22 and is
PASS: event.offsetY should be 104 and is
PASS: event.x should be 22 and is
PASS: event.y should be 104 and is
PASS: event type should be mousedown and is
PASS: event target should be [object HTMLOptionElement] and is
PASS: event.pageX should be 22 and is
PASS: event.pageY should be 104 and is
PASS: event.offsetX should be 22 and is
PASS: event.offsetY should be 104 and is
PASS: event.x should be 22 and is
PASS: event.y should be 104 and is
PASS: event type should be mouseup and is
PASS: event target should be [object HTMLOptionElement] and is
PASS: event.pageX should be 22 and is
PASS: event.pageY should be 184 and is
PASS: event.offsetX should be 22 and is
PASS: event.offsetY should be 184 and is
PASS: event.x should be 22 and is
PASS: event.y should be 184 and is
PASS: event type should be mouseup and is
PASS: event target should be [object HTMLOptionElement] and is
PASS: event.pageX should be 22 and is
PASS: event.pageY should be 184 and is
PASS: event.offsetX should be 22 and is
PASS: event.offsetY should be 184 and is
PASS: event.x should be 22 and is
PASS: event.y should be 184 and is
PASS: event type should be click and is
PASS: event target should be [object HTMLOptionElement] and is
PASS: event.pageX should be 22 and is
PASS: event.pageY should be 262 and is
PASS: event.offsetX should be 22 and is
PASS: event.offsetY should be 262 and is
PASS: event.x should be 22 and is
PASS: event.y should be 262 and is
PASS: event type should be click and is
PASS: event target should be [object HTMLOptionElement] and is
PASS: event.pageX should be 22 and is
PASS: event.pageY should be 262 and is
PASS: event.offsetX should be 22 and is
PASS: event.offsetY should be 262 and is
PASS: event.x should be 22 and is
PASS: event.y should be 262 and is
PASS: event type should be dblclick and is
PASS: event target should be [object HTMLOptionElement] and is
PASS: event.pageX should be 22 and is
PASS: event.pageY should be 344 and is
PASS: event.offsetX should be 22 and is
PASS: event.offsetY should be 344 and is
PASS: event.x should be 22 and is
PASS: event.y should be 344 and is
PASS: event type should be dblclick and is
PASS: event target should be [object HTMLOptionElement] and is
PASS: event.pageX should be 22 and is
PASS: event.pageY should be 344 and is
PASS: event.offsetX should be 22 and is
PASS: event.offsetY should be 344 and is
PASS: event.x should be 22 and is
PASS: event.y should be 344 and is
PASS: event type should be mousedown and is
PASS: event target should be [object HTMLSelectElement] and is
PASS: event.pageX should be 22 and is
PASS: event.pageY should be 448 and is
PASS: event.offsetX should be 12 and is
PASS: event.offsetY should be 38 and is
PASS: event.x should be 22 and is
PASS: event.y should be 448 and is
PASS: event type should be mousedown and is
PASS: event target should be [object HTMLSelectElement] and is
PASS: event.pageX should be 22 and is
PASS: event.pageY should be 448 and is
PASS: event.offsetX should be 12 and is
PASS: event.offsetY should be 38 and is
PASS: event.x should be 22 and is
PASS: event.y should be 448 and is
